Se cambio el sistema de build de tasks.json a CMake.

+++ ex2/include/Circulo.hpp
@@ -5,13 +5,13 @@
 class Circulo {
 private:
     Punto m_centro;
-    float m_radio;
+    double m_radio;
 public:
-    float getRadio() const;
-    void setRadio(float t_radio);
+    double getRadio() const;
+    void setRadio(double t_radio);
 
     const Punto& getCentro() const;
     void setCentro(const Punto& t_centro);
     
-    Circulo(Punto t_centro=Punto{}, float t_radio={});
+    Circulo(Punto t_centro=Punto{}, double t_radio={});
 };
