+++ ex2/include/Elipse.hpp
@@ -5,16 +5,16 @@
 class Elipse {
 private:
     Punto m_centro;
-    float m_mayor, m_menor;
+    double m_mayor, m_menor;
 public:
-    float getMayor() const;
-    void setMayor(float t_mayor);
+    double getMayor() const;
+    void setMayor(double t_mayor);
     
-    float getMenor() const;
-    void setMenor(float t_menor);
+    double getMenor() const;
+    void setMenor(double t_menor);
 
     const Punto& getCentro() const;
     void setCentro(const Punto& t_centro);
 
-    Elipse(Punto t_centro={}, float t_mayor={}, float t_menor={});
+    Elipse(Punto t_centro={}, double t_mayor={}, double t_menor={});
 };
