+++ ex2/src/Elipse.cpp
@@ -1,18 +1,18 @@
 #include "Elipse.hpp"
 
-float Elipse::getMayor() const {
+double Elipse::getMayor() const {
     return m_mayor;
 }
 
-void Elipse::setMayor(float t_mayor) {
+void Elipse::setMayor(double t_mayor) {
     m_mayor = t_mayor;
 }
 
-float Elipse::getMenor() const {
+double Elipse::getMenor() const {
     return m_menor;
 }
 
-void Elipse::setMenor(float t_menor) {
+void Elipse::setMenor(double t_menor) {
     m_menor = t_menor;
 }
 
@@ -24,5 +24,5 @@ void Elipse::setCentro(const Punto& t_centro) {
     m_centro = t_centro;
 }
 
-Elipse::Elipse(Punto t_centro, float t_mayor, float t_menor)
+Elipse::Elipse(Punto t_centro, double t_mayor, double t_menor)
     : m_centro(t_centro), m_mayor(t_mayor), m_menor(t_menor) {}
