+++ ex2/src/Punto.cpp
@@ -1,22 +1,22 @@
 #include "Punto.hpp"
 
-float Punto::getX() const {
+double Punto::getX() const {
     return m_x;
 }
 
-void Punto::setX(float t_x) {
+void Punto::setX(double t_x) {
     m_x = t_x;
 }
 
-float Punto::getY() const {
+double Punto::getY() const {
     return m_y;
 }
 
-void Punto::setY(float t_y) {
+void Punto::setY(double t_y) {
     m_y = t_y;
 }
 
-Punto::Punto(float t_x, float t_y) {
+Punto::Punto(double t_x, double t_y) {
     m_x = t_x;
     m_y = t_y;
 }
