+++ ex2/src/Rectangulo.cpp
@@ -8,21 +8,21 @@ void Rectangulo::setEsquina(const Punto& t_esquina) {
     m_esquina = t_esquina;
 }
 
-float Rectangulo::getBase() const {
+double Rectangulo::getBase() const {
     return m_base;
 }
 
-void Rectangulo::setBase(float t_base) {
+void Rectangulo::setBase(double t_base) {
     m_base = t_base;
 }
 
-float Rectangulo::getAltura() const {
+double Rectangulo::getAltura() const {
     return m_altura;
 }
 
-void Rectangulo::setAltura(float t_altura) {
+void Rectangulo::setAltura(double t_altura) {
     m_altura = t_altura;
 }
 
-Rectangulo::Rectangulo(Punto t_esquina, float t_base, float t_altura) 
+Rectangulo::Rectangulo(Punto t_esquina, double t_base, double t_altura) 
     : m_esquina(t_esquina), m_base(t_base), m_altura(t_altura) {}
